requirements.txt
pyserial==3.4
typer==0.9.0
yaspin==3.0.1
inquirer==3.2.4
flet==0.20.2